python/templates/requirements.txt
pycryptodome
flask~=3.0.2